Fix opt out cache tests
