+++ tests/core/test_opt_out_cache.py
@@ -1,26 +1,13 @@
-# -*- coding: utf-8 -*-
+import pytest
 
-from pytest import fixture
-
-from openfisca_core.simulation_builder import SimulationBuilder
-from openfisca_country_template import CountryTaxBenefitSystem
 from openfisca_country_template.entities import Person
-from openfisca_core.variables import Variable
+
+from openfisca_core import periods
 from openfisca_core.periods import MONTH
+from openfisca_core.variables import Variable
 
 
-@fixture
-def month():
-    return '2016-05'
-
-
-@fixture
-def make_isolated_simulation(month):
-    def _make_simulation(tbs, data):
-        builder = SimulationBuilder()
-        builder.default_period = month
-        return builder.build_from_variables(tbs, data)
-    return _make_simulation
+PERIOD = periods.period("2016-01")
 
 
 class input(Variable):
@@ -50,43 +37,34 @@ class output(Variable):
         return person('intermediate', period)
 
 
-def get_filled_tbs():
-    tax_benefit_system = CountryTaxBenefitSystem()
+@pytest.fixture(scope = "module", autouse = True)
+def add_variables_to_tax_benefit_system(tax_benefit_system):
     tax_benefit_system.add_variables(input, intermediate, output)
 
-    return tax_benefit_system
+
+@pytest.fixture(scope = "module", autouse = True)
+def add_variables_to_cache_blakclist(tax_benefit_system):
+    tax_benefit_system.cache_blacklist = set(['intermediate'])
 
 
-# TaxBenefitSystem instance declared after formulas
-
-
-tax_benefit_system = get_filled_tbs()
-
-
-tax_benefit_system.cache_blacklist = set(['intermediate'])
-
-
-def test_without_cache_opt_out(make_isolated_simulation, month):
-    simulation = make_isolated_simulation(tax_benefit_system, {'input': 1})
-    simulation.calculate('output', period = month)
+@pytest.mark.parametrize("simulation", [({'input': 1}, PERIOD)], indirect = True)
+def test_without_cache_opt_out(simulation):
+    simulation.calculate('output', period = PERIOD)
     intermediate_cache = simulation.persons.get_holder('intermediate')
-    assert(intermediate_cache.get_array(month) is not None)
+    assert(intermediate_cache.get_array(PERIOD) is not None)
 
 
-def test_with_cache_opt_out(make_isolated_simulation, month):
-    simulation = make_isolated_simulation(tax_benefit_system, {'input': 1})
+@pytest.mark.parametrize("simulation", [({'input': 1}, PERIOD)], indirect = True)
+def test_with_cache_opt_out(simulation):
     simulation.debug = True
     simulation.opt_out_cache = True
-    simulation.calculate('output', period = month)
+    simulation.calculate('output', period = PERIOD)
     intermediate_cache = simulation.persons.get_holder('intermediate')
-    assert(intermediate_cache.get_array(month) is None)
+    assert(intermediate_cache.get_array(PERIOD) is None)
 
 
-tax_benefit_system2 = get_filled_tbs()
-
-
-def test_with_no_blacklist(make_isolated_simulation, month):
-    simulation = make_isolated_simulation(tax_benefit_system2, {'input': 1})
-    simulation.calculate('output', period = month)
+@pytest.mark.parametrize("simulation", [({'input': 1}, PERIOD)], indirect = True)
+def test_with_no_blacklist(simulation):
+    simulation.calculate('output', period = PERIOD)
     intermediate_cache = simulation.persons.get_holder('intermediate')
-    assert(intermediate_cache.get_array(month) is not None)
+    assert(intermediate_cache.get_array(PERIOD) is not None)
